NetFR C API implementation
This is intended to (finally) be the chosen implementation type which will be
used for Looking Glass integration. It intends to support, via Libfabric,
messaging and RDMA write (host->client only) operations over TCP and RDMA.

This implementation uses Libfabric MSG endpoints rather than RDM endpoints due
to the additional complexity of implementing reliable datagram communications,
especially with regard to connection setup and detection of failed connections.
TCM is no longer required for connection setup.

Metadata, such as those for frame and cursor data, will not be included as part
of the API. Instead, the upper layer code should take care of interpreting the
received payloads, using the existing KVMFR codebase. This should reduce the
implementation complexity on both sides and keep the data transfer subsystem
flexible, similar to that of LGMP.

This version is only tested to a minimal degree, and does not yet have a
thread-safe context manager implementation. Further bugfixes will be included
when more implementation issues are found, and debugging options will also be
disabled by default when there's an initial fully working version of the system.

System documentation is in progress.